make new panel for correct display in the application window UI
